Move service creation into the service package so it can be re-used by Caddy.
